fix(test): use .Items for ContainerListResult len and uint16 for port
ContainerList now returns a struct, not a slice — use .Items for len().
Port.Num() returns uint16, not int — match type in assertion.

+++ udp_port_binding_test.go
@@ -54,7 +54,7 @@ func TestUDPPortBinding(t *testing.T) {
 
 		portNum := mappedPort.Num()
 		assert.Positive(t, portNum, "Port number should be greater than 0")
-		assert.LessOrEqual(t, portNum, 65535, "Port number should be valid UDP port range")
+		assert.LessOrEqual(t, portNum, uint16(65535), "Port number should be valid UDP port range")
 
 		// Verify the port is actually accessible (basic connectivity test)
 		hostIP, err := container.Host(ctx)
